+++ src/main/java/org/folio/mr/service/impl/MediatedRequestActionsServiceImpl.java
@@ -1,9 +1,11 @@
 package org.folio.mr.service.impl;
 
 import static java.lang.String.format;
+import static java.util.Optional.ofNullable;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.CLOSED_CANCELLED;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.CLOSED_DECLINED;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.CLOSED_FILLED;
+import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.OPEN_AWAITING_DELIVERY;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.OPEN_AWAITING_PICKUP;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.OPEN_IN_TRANSIT_FOR_APPROVAL;
 import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.OPEN_IN_TRANSIT_TO_BE_CHECKED_OUT;
@@ -11,6 +13,7 @@ import static org.folio.mr.domain.dto.MediatedRequest.StatusEnum.OPEN_ITEM_ARRIV
 import static org.folio.mr.domain.dto.Request.StatusEnum.OPEN_NOT_YET_FILLED;
 import static org.folio.mr.support.ConversionUtils.asString;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -24,6 +27,8 @@ import org.folio.mr.domain.dto.MediatedRequest;
 import org.folio.mr.domain.dto.Request;
 import org.folio.mr.domain.dto.RequestDeliveryAddress;
 import org.folio.mr.domain.dto.RequestPickupServicePoint;
+import org.folio.mr.domain.dto.User;
+import org.folio.mr.domain.dto.UserPersonal;
 import org.folio.mr.domain.entity.MediatedRequestEntity;
 import org.folio.mr.domain.entity.MediatedRequestStep;
 import org.folio.mr.domain.entity.MediatedRequestWorkflow;
@@ -37,6 +42,7 @@ import org.folio.mr.service.EcsRequestService;
 import org.folio.mr.service.InventoryService;
 import org.folio.mr.service.MediatedRequestActionsService;
 import org.folio.mr.service.SearchService;
+import org.folio.mr.service.UserService;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.springframework.stereotype.Service;
@@ -50,6 +56,7 @@ import lombok.extern.log4j.Log4j2;
 public class MediatedRequestActionsServiceImpl implements MediatedRequestActionsService {
   private final MediatedRequestsRepository mediatedRequestsRepository;
   private final InventoryService inventoryService;
+  private final UserService userService;
   private final MediatedRequestMapper mediatedRequestMapper;
   private final MediatedRequestWorkflowLogRepository workflowLogRepository;
   private final CirculationRequestService circulationRequestService;
@@ -75,25 +82,20 @@ public class MediatedRequestActionsServiceImpl implements MediatedRequestActions
   }
 
   private Request createLocalRequest(MediatedRequestEntity mediatedRequest) {
-    Request localRequest = circulationRequestService.create(mediatedRequest);
-    updateLocalRequest(localRequest);
-    return localRequest;
-  }
-
-  private void updateLocalRequest(Request request) {
-    log.info("updateLocalRequest:: updating local request {}", request::getId);
-    circulationRequestService.update(request);
+    return circulationRequestService.create(mediatedRequest);
   }
 
   private Request createEcsTlr(MediatedRequestEntity mediatedRequest) {
     EcsTlr ecsTlr = ecsRequestService.create(mediatedRequest);
     Request primaryRequest = circulationRequestService.get(ecsTlr.getPrimaryRequestId());
-    revertPrimaryRequestDeliveryInfo(mediatedRequest, primaryRequest);
+    revertPrimaryRequestRequesterInfo(mediatedRequest, primaryRequest);
     return primaryRequest;
   }
 
-  private void revertPrimaryRequestDeliveryInfo(MediatedRequestEntity mediatedRequest, Request primaryRequest) {
-    log.info("updatePrimaryRequest:: updating primary request {}", primaryRequest::getId);
+  private void revertPrimaryRequestRequesterInfo(MediatedRequestEntity mediatedRequest,
+    Request primaryRequest) {
+
+    log.info("revertPrimaryRequestRequesterInfo:: primary request ID {}", primaryRequest::getId);
     // Changing requesterId from fake proxy ID back to the real ID of the secure patron
     primaryRequest.setRequesterId(mediatedRequest.getRequesterId().toString());
     circulationRequestService.update(primaryRequest);
@@ -160,47 +162,74 @@ public class MediatedRequestActionsServiceImpl implements MediatedRequestActions
   @Override
   public MediatedRequest confirmItemArrival(String itemBarcode) {
     log.info("confirmItemArrival:: item barcode: {}", itemBarcode);
-    MediatedRequestEntity entity = findMediatedRequestForItemArrival(itemBarcode);
-    changeMediatedRequestStatus(entity, OPEN_ITEM_ARRIVED);
-    mediatedRequestsRepository.save(entity);
-    MediatedRequest dto = mediatedRequestMapper.mapEntityToDto(entity);
-    MediatedRequestContext context = new MediatedRequestContext(dto);
+    MediatedRequestEntity mediatedRequestEntity = findMediatedRequestForItemArrival(itemBarcode);
+    changeMediatedRequestStatus(mediatedRequestEntity, OPEN_ITEM_ARRIVED);
+    mediatedRequestsRepository.save(mediatedRequestEntity);
+    MediatedRequest mediatedRequestDto = mediatedRequestMapper.mapEntityToDto(mediatedRequestEntity);
+    MediatedRequestContext context = new MediatedRequestContext(mediatedRequestDto);
     findItem(context);
+    findRequester(context);
     extendMediatedRequest(context);
-    revertPrimaryRequestDeliveryInfo(dto);
+    revertConfirmedRequestDeliveryInfo(context);
 
-    log.debug("confirmItemArrival:: result: {}", dto);
-    return dto;
+    log.debug("confirmItemArrival:: result: {}", mediatedRequestDto);
+    return mediatedRequestDto;
   }
 
-  private void revertPrimaryRequestDeliveryInfo(MediatedRequest medRequest) {
-    log.info("revertPrimaryRequestDeliveryInfo:: medRequest: {}", medRequest.getId());
-    var primaryRequest = circulationRequestService.get(medRequest.getConfirmedRequestId());
-    primaryRequest.setFulfillmentPreference(Request.FulfillmentPreferenceEnum.fromValue(
-      medRequest.getFulfillmentPreference().getValue()));
-    var deliveryAddress = medRequest.getDeliveryAddress();
-    if (deliveryAddress != null) {
-      log.info("revertPrimaryRequestDeliveryInfo:: updating deliveryAddress for request: {}", medRequest.getId());
-      primaryRequest.setDeliveryAddress(new RequestDeliveryAddress()
-        .region(deliveryAddress.getRegion())
-        .city(deliveryAddress.getCity())
-        .countryId(deliveryAddress.getCountryId())
-        .addressTypeId(deliveryAddress.getAddressTypeId())
-        .addressLine1(deliveryAddress.getAddressLine1())
-        .addressLine2(deliveryAddress.getAddressLine2())
-        .postalCode(deliveryAddress.getPostalCode()));
+  private void revertConfirmedRequestDeliveryInfo(MediatedRequestContext context) {
+    var mediatedRequest = context.getRequest();
+    log.info("revertConfirmedRequestDeliveryInfo:: mediatedRequest: {}", mediatedRequest.getId());
+
+    // Reverting fulfillment preference
+    var confirmedRequest = circulationRequestService.get(mediatedRequest.getConfirmedRequestId());
+    confirmedRequest.setFulfillmentPreference(Request.FulfillmentPreferenceEnum.fromValue(
+      mediatedRequest.getFulfillmentPreference().getValue()));
+
+    // Reverting delivery address info
+    var deliveryAddressTypeId = mediatedRequest.getDeliveryAddressTypeId();
+    if (deliveryAddressTypeId != null) {
+      log.info("revertConfirmedRequestDeliveryInfo:: " +
+          "updating deliveryAddressTypeId; confirmed request: {}, mediated request: {}",
+        confirmedRequest::getId, mediatedRequest::getId);
+      confirmedRequest.setDeliveryAddressTypeId(deliveryAddressTypeId);
+
+      var deliveryAddress = ofNullable(context.getRequester())
+        .map(User::getPersonal)
+        .map(UserPersonal::getAddresses)
+        .map(Collection::stream)
+        .flatMap(addresses -> addresses
+          .filter(address -> deliveryAddressTypeId.equals(address.getAddressTypeId()))
+          .findFirst())
+        .map(address -> new RequestDeliveryAddress()
+          .addressLine1(address.getAddressLine1())
+          .addressLine2(address.getAddressLine2())
+          .city(address.getCity())
+          .postalCode(address.getPostalCode())
+          .region(address.getRegion())
+          .countryId(address.getCountryId()));
+
+      if (deliveryAddress.isPresent()) {
+        log.info("revertConfirmedRequestDeliveryInfo:: " +
+            "updating deliveryAddress; confirmed request: {}, mediated request: {}",
+          confirmedRequest::getId, mediatedRequest::getId);
+        confirmedRequest.setDeliveryAddress(deliveryAddress.get());
+      }
     }
-    primaryRequest.setPickupServicePointId(medRequest.getPickupServicePointId());
-    var medRequestPickupServicePoint = medRequest.getPickupServicePoint();
-    if (medRequestPickupServicePoint != null) {
-      log.info("revertPrimaryRequestDeliveryInfo:: updating pickupServicePoint for primary request: {}", medRequest.getId());
-      primaryRequest.setPickupServicePoint(new RequestPickupServicePoint()
-        .name(medRequestPickupServicePoint.getName())
-        .code(medRequestPickupServicePoint.getCode())
-        .discoveryDisplayName(medRequestPickupServicePoint.getDiscoveryDisplayName())
-        .pickupLocation(medRequestPickupServicePoint.getPickupLocation()));
+
+    // Reverting pickup service point
+    confirmedRequest.setPickupServicePointId(mediatedRequest.getPickupServicePointId());
+    var mediatedRequestPickupServicePoint = mediatedRequest.getPickupServicePoint();
+    if (mediatedRequestPickupServicePoint != null) {
+      log.info("revertConfirmedRequestDeliveryInfo:: " +
+          "updating pickupServicePoint; confirmed request: {}, mediated request: {}",
+        confirmedRequest::getId, mediatedRequest::getId);
+      confirmedRequest.setPickupServicePoint(new RequestPickupServicePoint()
+        .name(mediatedRequestPickupServicePoint.getName())
+        .code(mediatedRequestPickupServicePoint.getCode())
+        .discoveryDisplayName(mediatedRequestPickupServicePoint.getDiscoveryDisplayName())
+        .pickupLocation(mediatedRequestPickupServicePoint.getPickupLocation()));
     }
-    circulationRequestService.update(primaryRequest);
+    circulationRequestService.update(confirmedRequest);
   }
 
   @Override
@@ -229,6 +258,7 @@ public class MediatedRequestActionsServiceImpl implements MediatedRequestActions
     var dto = mediatedRequestMapper.mapEntityToDto(entity);
     MediatedRequestContext context = new MediatedRequestContext(dto);
     findItem(context);
+    findRequester(context);
     extendMediatedRequest(context);
 
     log.debug("sendItemInTransit:: result: {}", dto);
@@ -254,6 +284,10 @@ public class MediatedRequestActionsServiceImpl implements MediatedRequestActions
       .map(ConsortiumItem::getTenantId)
       .map(context::setLendingTenantId)
       .ifPresent(this::fetchItem);
+  }
+
+  private void findRequester(MediatedRequestContext context) {
+    context.setRequester(userService.fetchUser(context.getRequest().getRequesterId()));
   }
 
   private void fetchItem(MediatedRequestContext context) {
@@ -360,6 +394,12 @@ public class MediatedRequestActionsServiceImpl implements MediatedRequestActions
   public void changeStatusToAwaitingPickup(MediatedRequestEntity request) {
     log.info("changeStatusToAwaitingPickup:: request id: {}", request.getId());
     changeMediatedRequestStatus(request, OPEN_AWAITING_PICKUP);
+  }
+
+  @Override
+  public void changeStatusToAwaitingDelivery(MediatedRequestEntity request) {
+    log.info("changeStatusToAwaitingDelivery:: request id: {}", request.getId());
+    changeMediatedRequestStatus(request, OPEN_AWAITING_DELIVERY);
   }
 
   @Override